WIP warn and fill positional selectors
Some ruminations about how we might do this. Scanning with a regexp
is a bit hacky but might be good enough. The "pos at the end" is the
most common I suspect and easy to fill by calling the selector with
a the positional removed and then a positional-like method afterwards.
If we need to do POS inside selectors it's more complicated, but we
could also just warn those cases and not try to fill.